requirements.txt
# === Python Proje Bağımlılıkları (ELE 495 - Tüm Sistem) ===

# Web Arayüzü
flask

# Ses Girdisi ve İşleme
sounddevice
scipy
numpy
playsound

# Motor ve GPIO Kontrolü
RPi.GPIO

# Seri Haberleşme
pyserial

# API İstekleri
requests
retry
python-dotenv

# Google AI ve Cloud
google-cloud-texttospeech
google-generativeai